+++ sample/sample1-simple-entity/entity/Post.ts
@@ -1,10 +1,12 @@
 import {Column, Entity} from "../../../src/index";
 import {PrimaryColumn} from "../../../src/decorator/columns/PrimaryColumn";
+import {Generated} from "../../../src/decorator/Generated";
 
 @Entity("sample01_post")
 export class Post {
 
-    @PrimaryColumn("integer", { generated: true })
+    @PrimaryColumn("integer")
+    @Generated()
     id: number;
 
     @Column()