+++ sample/sample13-everywhere-abstraction/entity/BaseObject.ts
@@ -3,10 +3,12 @@ import {BasePost} from "./BasePost";
 import {PostAuthor} from "./PostAuthor";
 import {ManyToOne} from "../../../src/decorator/relations/ManyToOne";
 import {PrimaryColumn} from "../../../src/decorator/columns/PrimaryColumn";
+import {Generated} from "../../../src/decorator/Generated";
 
 export class BaseObject extends BasePost {
 
-    @PrimaryColumn("double", { generated: true })
+    @PrimaryColumn("double")
+    @Generated()
     id: number;
 
     @Column()