+++ src/decorator/columns/PrimaryGeneratedColumn.ts
@@ -3,6 +3,7 @@ import {ColumnMetadataArgs} from "../../metadata-args/ColumnMetadataArgs";
 import {PrimaryGeneratedColumnNumericOptions} from "../options/PrimaryGeneratedColumnNumericOptions";
 import {ColumnOptions} from "../options/ColumnOptions";
 import {PrimaryGeneratedColumnUUIDOptions} from "../options/PrimaryGeneratedColumnUUIDOptions";
+import {GeneratedMetadataArgs} from "../../metadata-args/GeneratedMetadataArgs";
 
 /**
  * Column decorator is used to mark a specific class property as a table column.
@@ -33,15 +34,18 @@ export function PrimaryGeneratedColumn(strategy: "uuid", options?: PrimaryGenera
 export function PrimaryGeneratedColumn(strategyOrOptions?: "increment"|"uuid"|PrimaryGeneratedColumnNumericOptions|PrimaryGeneratedColumnUUIDOptions,
                                        maybeOptions?: PrimaryGeneratedColumnNumericOptions|PrimaryGeneratedColumnUUIDOptions): Function {
     const options: ColumnOptions = {};
+    let strategy: "increment"|"uuid";
 
     if (strategyOrOptions) {
         if (typeof strategyOrOptions === "string")
-            options.generationStrategy = strategyOrOptions as "increment"|"uuid";
+            strategy = strategyOrOptions as "increment"|"uuid";
 
-        if (strategyOrOptions instanceof Object)
+        if (strategyOrOptions instanceof Object) {
+            strategy = "increment";
             Object.assign(options, strategyOrOptions);
+        }
     } else {
-        options.generationStrategy = "increment";
+        strategy = "increment";
     }
 
     if (maybeOptions instanceof Object)
@@ -51,7 +55,7 @@ export function PrimaryGeneratedColumn(strategyOrOptions?: "increment"|"uuid"|Pr
 
         // check if there is no type in column options then set the int type - by default for auto generated column
         if (!options.type) {
-            if (options.generationStrategy === "increment") {
+            if (strategy === "increment") {
                 Object.assign(options, { type: Number}  as ColumnOptions);
             } else {
                 Object.assign(options, { type: "uuid"}  as ColumnOptions);
@@ -59,16 +63,23 @@ export function PrimaryGeneratedColumn(strategyOrOptions?: "increment"|"uuid"|Pr
         }
 
         // implicitly set a primary and generated to column options
-        Object.assign(options, { primary: true, generated: true } as ColumnOptions);
+        Object.assign(options, { primary: true } as ColumnOptions);
 
         // create and register a new column metadata
-        const args: ColumnMetadataArgs = {
+        const columnArgs: ColumnMetadataArgs = {
             target: object.constructor,
             propertyName: propertyName,
             mode: "regular",
             options: options
         };
-        getMetadataArgsStorage().columns.push(args);
+        getMetadataArgsStorage().columns.push(columnArgs);
+
+        const generationArgs: GeneratedMetadataArgs = {
+            target: object.constructor,
+            propertyName: propertyName,
+            strategy: strategy
+        };
+        getMetadataArgsStorage().generations.push(generationArgs);
     };
 }