+++ src/metadata/ColumnMetadata.ts
@@ -216,10 +216,6 @@ export class ColumnMetadata {
             this.collation = options.args.options.collation;
         if (options.args.options.primary)
             this.isPrimary = options.args.options.primary;
-        if (options.args.options.generated)
-            this.isGenerated = options.args.options.generated;
-        if (options.args.options.generationStrategy)
-            this.generationStrategy = options.args.options.generationStrategy;
         if (options.args.options.unique)
             this.isUnique = options.args.options.unique;
         if (options.args.options.default === null) // to make sure default: null is the same as nullable: true