+++ test/functional/persistence/custom-column-name-pk/entity/Post.ts
@@ -3,11 +3,13 @@ import {PrimaryColumn} from "../../../../../src/decorator/columns/PrimaryColumn"
 import {Column} from "../../../../../src/decorator/columns/Column";
 import {ManyToOne} from "../../../../../src/decorator/relations/ManyToOne";
 import {Category} from "./Category";
+import {Generated} from "../../../../../src/decorator/Generated";
 
 @Entity()
 export class Post {
 
-    @PrimaryColumn("int", {generated: true, name: "theId"})
+    @PrimaryColumn("int", {name: "theId"})
+    @Generated()
     id: number;
 
     @Column()