+++ test/functional/query-builder/relation-id/one-to-many/embedded/embedded.ts
@@ -12,7 +12,7 @@ import {Counters} from "./entity/Counters";
 import {User} from "./entity/User";
 import {Subcounters} from "./entity/Subcounters";
 
-describe("query builder > relation-id > many-to-many > embedded", () => {
+describe("query builder > relation-id > one-to-many > embedded", () => {
 
     let connections: Connection[];
     before(async () => connections = await createTestingConnections({